requirements.txt
streamlit==1.43.2
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
python-dotenv==1.0.1
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-api-python-client==2.165.0
google-auth-oauthlib==1.2.1
httpx==0.28.1
openpyxl==3.1.5
